requirements.txt
python-telegram-bot==21.9
requests
Pillow
python-dotenv
PyMuPDF
